fix(tauri): redirect actions.ts to actions.client.ts dynamically using pre/post build scripts to prevent Turbopack compilation bypasses
